# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nexus AI Workspace API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: workspaces
    description: Workspace management
  - name: pages
    description: Page management
  - name: tasks
    description: Task management
  - name: snippets
    description: Code snippet management
  - name: sessions
    description: Focus session management
  - name: dashboard
    description: Dashboard and stats
  - name: anthropic
    description: Anthropic AI chat operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # Workspaces
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List all workspaces for the user
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a new workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceBody"
      responses:
        "201":
          description: Created workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"

  /workspaces/{id}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get a workspace by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateWorkspace
      tags: [workspaces]
      summary: Update a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceBody"
      responses:
        "200":
          description: Updated workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
    delete:
      operationId: deleteWorkspace
      tags: [workspaces]
      summary: Delete a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Pages
  /workspaces/{workspaceId}/pages:
    get:
      operationId: listPages
      tags: [pages]
      summary: List pages in a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Page"
    post:
      operationId: createPage
      tags: [pages]
      summary: Create a new page
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePageBody"
      responses:
        "201":
          description: Created page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"

  /workspaces/{workspaceId}/pages/{id}:
    get:
      operationId: getPage
      tags: [pages]
      summary: Get a page by id
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
    put:
      operationId: updatePage
      tags: [pages]
      summary: Update a page
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePageBody"
      responses:
        "200":
          description: Updated page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
    delete:
      operationId: deletePage
      tags: [pages]
      summary: Delete a page
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Tasks
  /workspaces/{workspaceId}/tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks in a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /workspaces/{workspaceId}/tasks/{id}:
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Snippets
  /snippets:
    get:
      operationId: listSnippets
      tags: [snippets]
      summary: List all snippets for the user
      responses:
        "200":
          description: List of snippets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Snippet"
    post:
      operationId: createSnippet
      tags: [snippets]
      summary: Create a new snippet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSnippetBody"
      responses:
        "201":
          description: Created snippet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snippet"

  /snippets/{id}:
    put:
      operationId: updateSnippet
      tags: [snippets]
      summary: Update a snippet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSnippetBody"
      responses:
        "200":
          description: Updated snippet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snippet"
    delete:
      operationId: deleteSnippet
      tags: [snippets]
      summary: Delete a snippet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Focus Sessions
  /sessions:
    get:
      operationId: listSessions
      tags: [sessions]
      summary: List focus sessions for the user
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FocusSession"
    post:
      operationId: createSession
      tags: [sessions]
      summary: Create a new focus session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionBody"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FocusSession"

  # Dashboard / Stats
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard stats for the current user
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/recent-pages:
    get:
      operationId: getRecentPages
      tags: [dashboard]
      summary: Get recently updated pages
      responses:
        "200":
          description: Recent pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecentPage"

  /dashboard/today-tasks:
    get:
      operationId: getTodayTasks
      tags: [dashboard]
      summary: Get tasks for today with priority
      responses:
        "200":
          description: Today's tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

  # Anthropic AI
  /anthropic/conversations:
    get:
      operationId: listAnthropicConversations
      tags: [anthropic]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicConversation"
    post:
      operationId: createAnthropicConversation
      tags: [anthropic]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnthropicConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversation"

  /anthropic/conversations/{id}:
    get:
      operationId: getAnthropicConversation
      tags: [anthropic]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"
    delete:
      operationId: deleteAnthropicConversation
      tags: [anthropic]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /anthropic/conversations/{id}/messages:
    get:
      operationId: listAnthropicMessages
      tags: [anthropic]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicMessage"
    post:
      operationId: sendAnthropicMessage
      tags: [anthropic]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendAnthropicMessageBody"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Workspace:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - createdAt
        - updatedAt

    CreateWorkspaceBody:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
      required:
        - name

    UpdateWorkspaceBody:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string

    Page:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        userId:
          type: string
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        priority:
          type: string
          enum: [high, medium, low]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - userId
        - title
        - content
        - createdAt
        - updatedAt

    CreatePageBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        priority:
          type: string
          enum: [high, medium, low]
      required:
        - title

    UpdatePageBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        priority:
          type: string
          enum: [high, medium, low]

    Task:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        userId:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [high, medium, low]
        deadline:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - userId
        - title
        - status
        - priority
        - createdAt
        - updatedAt

    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [high, medium, low]
        deadline:
          type: string
          format: date-time
      required:
        - title

    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [high, medium, low]
        deadline:
          type: string
          format: date-time

    Snippet:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        language:
          type: string
        code:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - language
        - code
        - description
        - createdAt
        - updatedAt

    CreateSnippetBody:
      type: object
      properties:
        language:
          type: string
        code:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - language
        - code
        - description

    UpdateSnippetBody:
      type: object
      properties:
        language:
          type: string
        code:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string

    FocusSession:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        duration:
          type: integer
        focusScore:
          type: integer
        tasksCompleted:
          type: integer
        date:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - duration
        - focusScore
        - tasksCompleted
        - date
        - createdAt

    CreateSessionBody:
      type: object
      properties:
        duration:
          type: integer
        focusScore:
          type: integer
        tasksCompleted:
          type: integer
        date:
          type: string
          format: date-time
      required:
        - duration
        - focusScore
        - tasksCompleted

    DashboardStats:
      type: object
      properties:
        tasksCompletedThisWeek:
          type: integer
        notesWritten:
          type: integer
        streakDays:
          type: integer
        flowScore:
          type: integer
        totalWorkspaces:
          type: integer
        totalTasks:
          type: integer
        totalPages:
          type: integer
      required:
        - tasksCompletedThisWeek
        - notesWritten
        - streakDays
        - flowScore
        - totalWorkspaces
        - totalTasks
        - totalPages

    RecentPage:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        workspaceId:
          type: integer
        workspaceName:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - workspaceId
        - workspaceName
        - updatedAt

    AnthropicConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    AnthropicMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    CreateAnthropicConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    SendAnthropicMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    AnthropicConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AnthropicMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    AnthropicError:
      type: object
      properties:
        error:
          type: string
      required:
        - error